backend/requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
deep-translator==1.11.4
apscheduler==3.10.4
playwright==1.42.0
httpx==0.27.0
duckduckgo-search==5.3.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
google-generativeai==0.4.1
python-multipart==0.0.9
itsdangerous==2.1.2
bcrypt==4.1.2
